ai-ml/requirements.txt
google-generativeai
groq
python-dotenv
requests
fastapi
uvicorn